Crate hyperactor

Hyperactor is an actor system intended for managing large scale compute.

Actor data model

Hyperactor is designed to support large scale (millions of nodes) machine learning workloads where actor topologies communicate through high fanout multicast messaging.

Supporting this scale requires us to impose additional structure at the level of the framework, so that we can efficiently refer to meshes of actors that implement the same worker runtimes.

Similarly, Hyperactor must co-schedule actors in order to support collective communicaton between actors.

Hyperactor is organized into a hierarchy:

Each proc represents a single actor runtime instance, and hosts zero or more actors.

Actors are spawned into procs, and assigned a global name. Actors spawned in this way are assigned a local PID (pid) of 0. Actors in turn can spawn local actors. These inherit the global pid of their parent, but receive a unique pid.

This scheme confers several benefits:

Routing of messages can be performed by prefix. For example, we can identify the proc of the actor and send the message to it, which can then in turn be routed locally.

We can represent meshes of actors in a uniform and compact way. This is the basis on which we implement efficient multicasting within the system.

Entity	Identifier

Proc	addr,proc_name

Actor	addr,proc_name,name[pid]

Re-exports

pub use actor::Actor;

pub use actor::ActorHandle;

pub use actor::Handler;

pub use actor::HandlerInfo;

pub use actor::RemoteHandles;

pub use actor::RemoteSpawn;

pub use actor_local::ActorLocal;

pub use id::Label;

pub use id::Uid;

pub use mailbox::Data;

pub use mailbox::Mailbox;

pub use mailbox::Message;

pub use mailbox::OncePortHandle;

pub use mailbox::PortHandle;

pub use mailbox::RemoteMessage;

pub use proc::Context;

pub use proc::Instance;

pub use proc::InstanceCell;

pub use proc::Proc;

pub use proc::WeakProc;

pub use remote::Accepts;

Modules

accum

Defines the accumulator trait and some common accumulators.

actor

This module contains all the core traits required to define and manage actors.

actor_local

Per-actor local storage, analogous to thread_local! but scoped to actor instances.

channel

One-way, multi-process, typed communication channels. These are used to send messages between mailboxes residing in different processes.

config

Configuration keys and I/O for hyperactor.

context

This module defines traits that are used as context arguments to various hyperactor APIs; usually crate::context::Actor, implemented by crate::proc::Context (provided to actor handlers) and crate::proc::Instance, representing a running actor instance.

host

This module defines Host, which represents all the procs running on a host. The procs themselves are managed by an implementation of ProcManager, which may, for example, fork new processes for each proc, or spawn them in the same process for testing purposes.

id

Universal identifier types for the actor system.

introspect

Introspection protocol for hyperactor actors.

mailbox

Mailboxes are the central message-passing mechanism in Hyperactor.

message

This module provides a framework for mutating serialized messages without the need to deserialize them. This capability is useful when sending messages to a remote destination throughout intermeidate nodes, where the intermediate nodes do not contain the message’s type information.

metrics

Hyperactor metrics.

ordering

This module contains utilities used to help messages are delivered in order for any given sender and receiver actor pair.

panic_handler

Used to capture the backtrace from panic and store it in a task_local, so that it can be retrieved later when the panic is catched.

port

Port identifier newtype.

proc

Proc is an addressable actor-runtime boundary.

ref_

References: identifiers paired with a network location.

reference

References for different resources in Hyperactor.

remote

Remote references: typed wrappers around ActorRef or PortRef for type-safe message sending.

supervision

Messages used in supervision.

sync

Synchronization primitives that are used by Hyperactor.

testing

Test utilities.

time

This module contains various utilities for dealing with time. (This probably belongs in a separate crate.)

Macros

assert_behaves

Check if the actor behaves-as the a given behavior (defined by [behavior!]).

assert_cancel_safe

Convenience macro for assert_cancel_safe.

assert_cancel_safe_async

Async convenience macro for assert_cancel_safe_async.

behavior

Create a [Referable] definition, handling a specific set of message types. Behaviors are used to create an [ActorRef] without having to depend on the actor’s implementation. If the message type need to be cast, add castable flag to those types. e.g. the following example creates a behavior with 5 message types, and 4 of which need to be cast.

declare_static_counter

Create a thread safe static counter that can be incremeneted or decremented. This is useful to avoid creating temporary counters. You can safely create counters with the same name. They will be joined by the underlying runtime and are thread safe.

declare_static_gauge

Create a thread safe static gauge that can be set to a specific value. This is useful to avoid creating temporary gauges. You can safely create gauges with the same name. They will be joined by the underlying runtime and are thread safe.

declare_static_histogram

Create a thread safe static histogram that can be incremeneted or decremented. This is useful to avoid creating temporary histograms. You can safely create histograms with the same name. They will be joined by the underlying runtime and are thread safe.

declare_static_timer

Create a thread safe static timer that can be used to measure durations. This macro creates a histogram with predefined boundaries appropriate for the specified time unit. Supported units are “ms” (milliseconds), “us” (microseconds), and “ns” (nanoseconds).

key_value

Create key value pairs for use in opentelemetry. These pairs can be stored and used multiple times. Opentelemetry adds key value attributes when you bump counters and histograms. so MY_COUNTER.add(42, &[key_value!(“key”, “value”)]) and MY_COUNTER.add(42, &[key_value!(“key”, “other_value”)]) will actually bump two separete counters.

kv_pairs

Construct the key value attribute slice using mapping syntax. Example:

remote

Register an actor type so that it can be spawned remotely. The actor type must implement typeuri::Named, which will be used to identify the actor globally.

uid

Compile-time validated [hyperactor::id::Uid] construction.

Structs

SignalCleanupGuard

RAII guard that automatically unregisters a signal cleanup callback when dropped

Enums

SignalDisposition

This type describes how a signal is currently handled by the process.

Functions

initialize

Initialize the Hyperactor runtime. Specifically:

initialize_with_current_runtime

Initialize the Hyperactor runtime using the current tokio runtime handle.

initialize_with_log_prefix

Initialize the Hyperactor runtime. Specifically:

query_signal_disposition

Query the current disposition of a signal (signum).

register_signal_cleanup

Register a cleanup callback to be executed on SIGINT/SIGTERM Returns a unique ID that can be used to unregister the callback

register_signal_cleanup_scoped

Register a scoped cleanup callback to be executed on SIGINT/SIGTERM Returns a guard that automatically unregisters the callback when dropped

sigpipe_disposition

Returns the current SignalDisposition of SIGPIPE.

unregister_signal_cleanup

Unregister a previously registered cleanup callback

Attribute Macros

export

Exports handlers for this actor. The set of exported handlers determine the messages that may be sent to remote references of the actor ([hyperaxtor::ActorRef]). Only messages that implement [hyperactor::RemoteMessage] may be exported.

handle

Install a Handler that routes messages of the provided type to this handler trait implementation.

instrument

Use this macro in place of tracing::instrument to prevent spamming our tracing table. We set a default level of INFO while always setting ERROR if the function returns Result::Err giving us consistent and high quality structured logs. Because this wraps around tracing::instrument, all parameters mentioned in https://fburl.com/9jlkb5q4 should be valid. For functions that don’t return a [Result] type, use [instrument_infallible]

instrument_infallible

Use this macro in place of tracing::instrument to prevent spamming our tracing table. Because this wraps around tracing::instrument, all parameters mentioned in https://fburl.com/9jlkb5q4 should be valid.

observe_async

A procedural macro that automatically injects telemetry code into async functions that do not return a Result type.

observe_result

A procedural macro that automatically injects telemetry code into async functions that return a Result type.

Derive Macros

Bind

Derive a custom implementation of [hyperactor::message::Bind] trait for a struct or enum. This macro is normally used in tandem with [fn derive_unbind] to make the applied struct or enum castable.

HandleClient

Derives a client implementation on ActorHandle<Actor>. See Handler documentation for details.

Handler

Derive a custom handler trait for given an enum containing tuple structs. The handler trait defines a method corresponding to each of the enum’s variants, and a handle function that dispatches messages to the correct method. The macro supports two messaging patterns: “call” and “oneway”. A call is a request-response message; a [hyperactor::mailbox::OncePortRef] or [hyperactor::mailbox::OncePortHandle] in the last position is used to send the return value.

RefClient

Derives a client implementation on ActorRef<Actor>. See Handler documentation for details.

Unbind

Derive a custom implementation of [hyperactor::message::Unbind] trait for a struct or enum. This macro is normally used in tandem with [fn derive_bind] to make the applied struct or enum castable.