hyperactor::accum

Enum ReducerMode

pub enum ReducerMode {
    Streaming(StreamingReducerOpts),
    Once(usize),
}

The mode in which a reducer operates.

Variants

Streaming(StreamingReducerOpts)

Streaming mode: continuously reduce and emit updates based on buffer size/timeout.

Once(usize)

Once mode: accumulate exactly n values, emit a single reduced update, then tear down.

Implementations

impl ReducerMode

pub fn is_streaming(&self) -> bool

Returns true if this is a ReducerMode::Streaming, otherwise false

pub fn as_streaming_mut(&mut self) -> Option<&mut StreamingReducerOpts>

Optionally returns mutable references to the inner fields if this is a ReducerMode::Streaming, otherwise None

pub fn as_streaming(&self) -> Option<&StreamingReducerOpts>

Optionally returns references to the inner fields if this is a ReducerMode::Streaming, otherwise None

pub fn into_streaming(self) -> Result<StreamingReducerOpts, Self>

Returns the inner fields if this is a ReducerMode::Streaming, otherwise returns back the enum in the Err case of the result

pub fn is_once(&self) -> bool

Returns true if this is a ReducerMode::Once, otherwise false

pub fn as_once_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ReducerMode::Once, otherwise None

pub fn as_once(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ReducerMode::Once, otherwise None

pub fn into_once(self) -> Result<usize, Self>

Returns the inner fields if this is a ReducerMode::Once, otherwise returns back the enum in the Err case of the result