hyperactor

Module actor

This module contains all the core traits required to define and manage actors.

Modules

remote

Management of actor registration for remote spawning.

Structs

ActorError

Errors that occur while serving actors. Each error is associated with the ID of the actor being served.

ActorHandle

ActorHandles represent a (local) serving actor. It is used to access its messaging and signal ports, as well as to synchronize with its lifecycle (e.g., providing joins). Once dropped, the handle is detached from the underlying actor instance, and there is no longer any way to join it.

HandlerInfo

Information about a message handler being processed.

Enums

ActorErrorKind

The kinds of actor serving errors.

ActorStatus

The runtime status of an actor.

Signal

A collection of signals to control the behavior of the actor. Signals are internal runtime control plane messages and should not be sent outside of the runtime.

Traits

Actor

An Actor is an independent, asynchronous thread of execution. Each actor instance has a mailbox, whose messages are delivered through the method [Actor::handle].

Binds

Binds determines how an actor’s ports are bound to a specific reference type.

Handler

A Handler allows an actor to handle a specific message type.

Referable

Referable is a marker trait for types that can appear as remote references across process boundaries.

RemoteHandles

Handles is a marker trait specifying that message type [M] is handled by a specific actor type.

RemoteSpawn

An Actor that can be spawned remotely.

Functions

handle_undeliverable_message

Default implementation of Actor::handle_undeliverable_message. Defined as a free function so that Actor implementations that override Actor::handle_undeliverable_message can fallback to this default.