hyperactor::actor

Enum ActorStatus

pub enum ActorStatus {
    Unknown,
    Created,
    Initializing,
    Client,
    Idle,
    Processing(SystemTime, Option<(String, Option<String>)>),
    Saving(SystemTime),
    Loading(SystemTime),
    Stopping,
    Stopped,
    Failed(ActorErrorKind),
}

The runtime status of an actor.

Variants

Unknown

The actor status is unknown.

Created

The actor was created, but not yet started.

Initializing

The actor is initializing. It is not yet ready to receive messages.

Client

The actor is in “client” state: the user is managing the actor’s mailboxes manually.

Idle

The actor is ready to receive messages, but is currently idle.

Processing(SystemTime, Option<(String, Option<String>)>)

The actor has been processing a message, beginning at the specified instant. The message handler and arm is included. TODO: we shoudl use interned representations here, so we don’t copy strings willy-nilly.

Saving(SystemTime)

The actor has been saving its state.

Loading(SystemTime)

The actor has been loading its state.

Stopping

The actor is stopping. It is draining messages.

Stopped

The actor is stopped. It is no longer processing messages.

Failed(ActorErrorKind)

The actor failed with the provided actor error.

Implementations

impl ActorStatus

pub fn is_unknown(&self) -> bool

Returns true if this is a ActorStatus::Unknown, otherwise false

pub fn is_created(&self) -> bool

Returns true if this is a ActorStatus::Created, otherwise false

pub fn is_initializing(&self) -> bool

Returns true if this is a ActorStatus::Initializing, otherwise false

pub fn is_client(&self) -> bool

Returns true if this is a ActorStatus::Client, otherwise false

pub fn is_idle(&self) -> bool

Returns true if this is a ActorStatus::Idle, otherwise false

pub fn is_processing(&self) -> bool

Returns true if this is a ActorStatus::Processing, otherwise false

pub fn as_processing_mut( &mut self, ) -> Option<(&mut SystemTime, &mut Option<(String, Option<String>)>)>

Optionally returns mutable references to the inner fields if this is a ActorStatus::Processing, otherwise None

pub fn as_processing( &self, ) -> Option<(&SystemTime, &Option<(String, Option<String>)>)>

Optionally returns references to the inner fields if this is a ActorStatus::Processing, otherwise None

pub fn into_processing( self, ) -> Result<(SystemTime, Option<(String, Option<String>)>), Self>

Returns the inner fields if this is a ActorStatus::Processing, otherwise returns back the enum in the Err case of the result

pub fn is_saving(&self) -> bool

Returns true if this is a ActorStatus::Saving, otherwise false

pub fn as_saving_mut(&mut self) -> Option<&mut SystemTime>

Optionally returns mutable references to the inner fields if this is a ActorStatus::Saving, otherwise None

pub fn as_saving(&self) -> Option<&SystemTime>

Optionally returns references to the inner fields if this is a ActorStatus::Saving, otherwise None

pub fn into_saving(self) -> Result<SystemTime, Self>

Returns the inner fields if this is a ActorStatus::Saving, otherwise returns back the enum in the Err case of the result

pub fn is_loading(&self) -> bool

Returns true if this is a ActorStatus::Loading, otherwise false

pub fn as_loading_mut(&mut self) -> Option<&mut SystemTime>

Optionally returns mutable references to the inner fields if this is a ActorStatus::Loading, otherwise None

pub fn as_loading(&self) -> Option<&SystemTime>

Optionally returns references to the inner fields if this is a ActorStatus::Loading, otherwise None

pub fn into_loading(self) -> Result<SystemTime, Self>

Returns the inner fields if this is a ActorStatus::Loading, otherwise returns back the enum in the Err case of the result

pub fn is_stopping(&self) -> bool

Returns true if this is a ActorStatus::Stopping, otherwise false

pub fn is_stopped(&self) -> bool

Returns true if this is a ActorStatus::Stopped, otherwise false

pub fn is_failed(&self) -> bool

Returns true if this is a ActorStatus::Failed, otherwise false

pub fn as_failed_mut(&mut self) -> Option<&mut ActorErrorKind>

Optionally returns mutable references to the inner fields if this is a ActorStatus::Failed, otherwise None

pub fn as_failed(&self) -> Option<&ActorErrorKind>

Optionally returns references to the inner fields if this is a ActorStatus::Failed, otherwise None

pub fn into_failed(self) -> Result<ActorErrorKind, Self>

Returns the inner fields if this is a ActorStatus::Failed, otherwise returns back the enum in the Err case of the result

impl ActorStatus

pub fn is_terminal(&self) -> bool

Tells whether the status is a terminal state.

pub fn generic_failure(message: impl Into<String>) -> Self

Create a generic failure status with the provided error message.