hyperactor::actor

Trait Actor

pub trait Actor:
    Sized
    + Send
    + 'static {
    // Provided methods
    fn init<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        _this: &'life1 Instance<Self>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn handle_stop<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        this: &'life1 Instance<Self>,
        mode: StopMode,
        reason: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn cleanup<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        _this: &'life1 Instance<Self>,
        _err: Option<&'life2 ActorError>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn spawn_server_task<F>(future: F) -> JoinHandle<F::Output>
       where F: Future + Send + 'static,
             F::Output: Send + 'static { ... }
    fn handle_supervision_event<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        _this: &'life1 Instance<Self>,
        event: &'life2 ActorSupervisionEvent,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn handle_delivery_failure_event<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        cx: &'life1 Instance<Self>,
        undeliverable: Undeliverable<MessageEnvelope>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn handle_undeliverable_message<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        cx: &'life1 Instance<Self>,
        reason: UndeliverableReason,
        undeliverable: Undeliverable<MessageEnvelope>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn handle_invalid_reference<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        cx: &'life1 Instance<Self>,
        invalid: InvalidReference,
        undeliverable: Undeliverable<MessageEnvelope>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn handle_expired_delivery<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        cx: &'life1 Instance<Self>,
        expired: ExpiredDelivery,
        undeliverable: Undeliverable<MessageEnvelope>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn display_name(&self) -> Option<String> { ... }
}

An Actor is an independent, asynchronous thread of execution. Each actor instance has a mailbox, whose messages are delivered through the method [Actor::handle].

Actors communicate with each other by way of message passing. Actors are assumed to be deterministic: that is, the state of an actor is determined by the set (and order) of messages it receives.

Provided Methods

fn init<'life0, 'life1, 'async_trait>( &'life0 mut self, _this: &'life1 Instance<Self>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Initialize the actor, after the runtime has been fully initialized. Init thus provides a mechanism by which an actor can reliably and always receive some initial event that can be used to kick off further (potentially delayed) processing.

fn handle_stop<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, this: &'life1 Instance<Self>, mode: StopMode, reason: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Handle a stop request from the runtime.

The default implementation closes handler ingress and then either exits immediately or queues an exit after already accepted handler work drains. Actors that need to coordinate asynchronous shutdown work can override this method and call Instance::exit() / Instance::exit_after_drain() later, once they are ready to terminate.

fn cleanup<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, _this: &'life1 Instance<Self>, _err: Option<&'life2 ActorError>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Cleanup things used by this actor before shutting down. Notably this function is async and allows more complex cleanup. Simpler cleanup can be handled by the impl Drop for this Actor. If err is not None, it is the error that this actor is failing with. Any errors returned by this function will be logged and ignored. If err is None, any errors returned by this function will be propagated as an ActorError. This function is not called if there is a panic in the actor, as the actor may be in an indeterminate state. It is also not called if the process is killed, there is no atexit handler or signal handler.

fn spawn_server_task<F>(future: F) -> JoinHandle<F::Output>
where F: Future + Send + 'static, F::Output: Send + 'static,

This method is used by the runtime to spawn the actor server. It can be used by actors that require customized runtime setups (e.g., dedicated actor threads), or want to use a custom tokio runtime.

fn handle_supervision_event<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, _this: &'life1 Instance<Self>, event: &'life2 ActorSupervisionEvent, ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Handle actor supervision event. Return `Ok(true)`` if the event is handled here.

fn handle_delivery_failure_event<'life0, 'life1, 'async_trait>( &'life0 mut self, cx: &'life1 Instance<Self>, undeliverable: Undeliverable<MessageEnvelope>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Default delivery-failure event handling behavior.

fn handle_undeliverable_message<'life0, 'life1, 'async_trait>( &'life0 mut self, cx: &'life1 Instance<Self>, reason: UndeliverableReason, undeliverable: Undeliverable<MessageEnvelope>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Default undeliverable message handling behavior.

fn handle_invalid_reference<'life0, 'life1, 'async_trait>( &'life0 mut self, cx: &'life1 Instance<Self>, invalid: InvalidReference, undeliverable: Undeliverable<MessageEnvelope>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Default invalid-reference handling behavior.

fn handle_expired_delivery<'life0, 'life1, 'async_trait>( &'life0 mut self, cx: &'life1 Instance<Self>, expired: ExpiredDelivery, undeliverable: Undeliverable<MessageEnvelope>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Default expired-delivery handling behavior.

fn display_name(&self) -> Option<String>

If overridden, we will use this name in place of the ActorAddr for talking about this actor in supervision error messages.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.