hyperactor

Module context

This module defines traits that are used as context arguments to various hyperactor APIs; usually crate::context::Actor, implemented by crate::proc::Context (provided to actor handlers) and crate::proc::Instance, representing a running actor instance.

Context traits are sealed, and thus can only be implemented by data types in the core hyperactor crate.

Traits

Actor

A typed actor context, providing both a Mailbox and an Instance.

Mailbox

A mailbox context provides a mailbox.