hyperactor::host

Trait ProcManager

pub trait ProcManager {
    type Handle: ProcHandle;
    type Config = ();

    // Required methods
    fn transport(&self) -> ChannelTransport;
    fn spawn<'life0, 'async_trait>(
        &'life0 self,
        proc_id: ProcId,
        forwarder_addr: ChannelAddr,
        config: Self::Config,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Handle, HostError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A trait describing a manager of procs, responsible for bootstrapping procs on a host, and managing their lifetimes. The manager spawns an Agent-typed actor on each proc, responsible for managing the proc.

Required Associated Types

type Handle: ProcHandle

Concrete handle type this manager returns.

Provided Associated Types

type Config = ()

Additional configuration for the proc, supported by this manager.

Required Methods

fn transport(&self) -> ChannelTransport

The preferred transport for this ProcManager. In practice this will be ChannelTransport::Local for testing, and ChannelTransport::Unix for external processes.

fn spawn<'life0, 'async_trait>( &'life0 self, proc_id: ProcId, forwarder_addr: ChannelAddr, config: Self::Config, ) -> Pin<Box<dyn Future<Output = Result<Self::Handle, HostError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Spawn a new proc with the provided proc id. The proc should use the provided forwarder address for messages destined outside of the proc. The returned address accepts messages destined for the proc.

An agent actor is also spawned, and the corresponding actor ref is returned.