hyperactor::mailbox::durable_mailbox_sender

Module log

This module implements a write-ahead log for mailboxes. This can be used to provide durable messaging facilities for actors.

Enums

MessageLogError

Errors that occur during message log operations. This enum is marked non-exhaustive to allow for extensibility.

Traits

MessageLog

This MessageLog is a log that serves as a building block to persist data before the consumer can process it. One typical example is to persist messages before an actor handles it. In such a case, it can be used as a white-ahead log. It allows the actor to recover from a crash without requesting resending the messages. The log is append-only and the messages are persisted in order with sequence ids.

Type Aliases

SeqId

A sequence id is a unique identifier for a message.