hyperactor::mailbox

Struct DialMailboxRouter

pub struct DialMailboxRouter { /* private fields */ }

A dynamic mailbox router that supports remote delivery.

DialMailboxRouter maintains a runtime address book mapping references to ChannelAddrs. It holds a cache of active connections and forwards messages to the appropriate MailboxClient.

If a message destination is not bound, but is a “direct mode” address (i.e., its proc id contains the channel address through which the proc is reachable), then DialMailboxRouter dials the proc directly.

Messages sent to unknown destinations are routed to the default sender, if present.

Implementations

impl DialMailboxRouter

pub fn new() -> Self

Create a new DialMailboxRouter with an empty routing table.

pub fn new_with_default(default: BoxedMailboxSender) -> Self

Create a new DialMailboxRouter with an empty routing table, and a default sender. Any message with an unknown destination is dispatched on this default sender, unless the destination is direct-addressed, in which case it is dialed directly.

pub fn new_with_default_direct_addressed_remote_only( default: BoxedMailboxSender, ) -> Self

Create a new DialMailboxRouter with an empty routing table, and a default sender. Any message with an unknown destination is dispatched on this default sender, unless the destination is direct-addressed and has a remote channel transport type.

pub fn bind(&self, dest: Reference, addr: ChannelAddr)

Binds a Reference to a ChannelAddr, replacing any existing binding.

If the address changes, the old sender is evicted from the cache to ensure fresh routing on next use.

pub fn unbind(&self, dest: &Reference)

Removes all address mappings with the given prefix from the router.

Also evicts any corresponding cached senders to prevent reuse of stale connections.

pub fn lookup_addr(&self, actor_id: &ActorId) -> Option<ChannelAddr>

Lookup an actor’s channel in the router’s address bok.

pub fn prefixes(&self) -> BTreeSet<Reference>

Return all covering prefixes of this router. That is, all references that are not prefixed by another reference in the routing table