hyperactor::mailbox

Struct MessageEnvelope

pub struct MessageEnvelope { /* private fields */ }

An envelope that carries a message destined to a remote actor. The envelope contains a serialized message along with its destination and sender.

Implementations

impl MessageEnvelope

pub fn new(sender: ActorId, dest: PortId, data: Any, headers: Attrs) -> Self

Create a new envelope with the provided sender, destination, and message.

pub fn serialize<T: Serialize + Named>( source: ActorId, dest: PortId, value: &T, headers: Attrs, ) -> Result<Self, Error>

Construct a new serialized value by serializing the provided T-typed value.

pub fn ttl(&self) -> u8

Returns the remaining time-to-live (TTL) for this message.

The TTL is decremented at each MailboxSender hop. When it reaches 0, the message is considered expired and is returned to the sender as undeliverable.

pub fn set_ttl(self, ttl: u8) -> Self

Overrides the message’s time-to-live (TTL).

This replaces the current TTL value (normally initialized from config::MESSAGE_TTL_DEFAULT) with the provided ttl. The updated envelope is returned for chaining.

Note

The TTL is decremented at each MailboxSender hop, and when it reaches 0 the message will be treated as undeliverable.

pub fn deserialized<T: DeserializeOwned + Named>(&self) -> Result<T, Error>

Deserialize the message in the envelope to the provided type T.

pub fn data(&self) -> &Any

The serialized message.

pub fn sender(&self) -> &ActorId

The message sender.

pub fn dest(&self) -> &PortId

The destination of the message.

pub fn headers(&self) -> &Attrs

The message headers.

pub fn is_signal(&self) -> bool

Tells whether this is a signal message.

pub fn set_error(&mut self, error: DeliveryError)

Set a delivery error for the message. If errors are already set, append it to the existing errors.

pub fn update_sender(&mut self, sender: ActorId)

Change the sender on the envelope in case it was set incorrectly. This should only be used by CommActor since it is forwarding from another sender.

pub fn set_return_undeliverable(&mut self, return_undeliverable: bool)

Set to true if you want this message to be returned to sender if it cannot reach dest. This is the default. Set to false if you want the message to be dropped instead.

pub fn undeliverable( self, error: DeliveryError, return_handle: PortHandle<Undeliverable<MessageEnvelope>>, )

The message has been determined to be undeliverable with the provided error. Mark the envelope with the error and return to sender.

pub fn errors(&self) -> &Vec<DeliveryError>

Get the errors of why this message was undeliverable. Empty means this message was not determined as undeliverable.

pub fn error_msg(&self) -> Option<String>

Get the string representation of the errors of this message was undeliverable. None means this message was not determined as undeliverable.

pub fn headers_mut(&mut self) -> &mut Attrs

Get a mutable reference to this envelope’s headers.