hyperactor::message

Trait Bind

pub trait Bind: Sized {
    // Required method
    fn bind(&mut self, bindings: &mut Bindings) -> Result<()>;
}

An object T that is Bind can bind a set of externally provided parameters into itself.

Required Methods

fn bind(&mut self, bindings: &mut Bindings) -> Result<()>

Remove parameters from bindings, and use them to update itself.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.