hyperactor::message

Trait Unbind

pub trait Unbind: Sized {
    // Required method
    fn unbind(&self, bindings: &mut Bindings) -> Result<()>;
}

An object T that is Unbind can extract a set of parameters from itself, and store in Bindings. The extracted parameters in Bindings can be independently manipulated, and then later reconstituted (rebound) into a T-typed object again.

Required Methods

fn unbind(&self, bindings: &mut Bindings) -> Result<()>

Extract parameters from itself and store them in bindings.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.