hyperactor

Module metrics

Hyperactor metrics.

This module contains metrics definitions for various components of hyperactor.

Enums

ChannelErrorType

Error types for channel-related errors. Only used for telemetry.

Constants

SERVER_HEARTBEAT_METRIC_NAME

Metric name for server heartbeat, used in both metric emission and client error messages.

Statics

ACTOR_MESSAGES_RECEIVED

a global counter named: actor.messages_received

ACTOR_MESSAGES_SENT

a global counter named: actor.messages_sent

ACTOR_MESSAGE_HANDLER_DURATION

a global histogram timer named: actor.message_handler_duration

ACTOR_MESSAGE_QUEUE_SIZE

a global up down counter named: actor.message_queue_size

ACTOR_MESSAGE_RECEIVE_ERRORS

a global counter named: actor.message_receive_errors

CHANNEL_CONNECTIONS

a global counter named: channel.connections

CHANNEL_ERRORS

a global counter named: channel.errors

CHANNEL_LATENCY_MICROS

a global histogram named: channel.latency.us

CHANNEL_NET_RX_BUFFER_FULL

a global counter named: channel.net_rx_buffer_full

CHANNEL_RECONNECTIONS

a global counter named: channel.reconnections

CHANNEL_THROUGHPUT_BYTES

a global counter named: channel.throughput.bytes

CHANNEL_THROUGHPUT_MESSAGES

a global counter named: channel.throughput.messages

MAILBOX_POSTS

a global counter named: mailbox.posts

MAILBOX_UNDELIVERABLE_MESSAGES

a global counter named: mailbox.undeliverable_messages

MESSAGE_LATENCY_MICROS

a global histogram named: message.e2e_latency.us

PROC_MESH_ACTOR_FAILURES

a global counter named: proc_mesh.actor_failures

PROC_MESH_ALLOCATION

a global counter named: proc_mesh.active_procs

PROC_MESH_PROC_STOPPED

a global counter named: proc_mesh.proc_failures

REMOTE_MESSAGE_SEND_SIZE

a global histogram named: channel.remote_message_send_size

SERVER_HEARTBEAT

a global counter named: channel.server.heartbeat