hyperactor

Module reference

References for different resources in Hyperactor.

The “Id” variants are transparent and typeless, whereas the corresponding “Ref” variants are opaque and typed. The latter intended to be exposed in user-facing APIs. We provide std::convert::From implementations between Id and Refs where this makes sense.

All system implementations use the same concrete reference representations, as their specific layout (e.g., actor index, rank, etc.) are used by the core communications algorithms throughout.

References and ids are crate::Messages to facilitate passing them between actors.

Modules

lex

This module contains a lexer for Hyperactor identifiers.

name

Names in hyperactor.

Structs

ActorId

Actors are identified by their proc, their name, and pid.

ActorRef

ActorRefs are typed references to actors.

OncePortRef

A remote reference to a [OncePort]. References are serializable and may be passed to remote actors, which can then use it to send a message to this port.

PortId

Port ids identify [crate::mailbox::Port]s of an actor.

PortRef

A reference to a remote port. All messages passed through PortRefs will be serialized. PortRefs are always streaming.

ProcId

Procs are identified by a direct channel address and local name. Each proc represents an actor runtime that can locally route to all of its constituent actors.

UnboundPort

The parameters extracted from PortRef to Bindings.

Enums

Reference

A universal reference to hierarchical identifiers in Hyperactor.

ReferenceKind

The kinds of references.

ReferenceParsingError

The type of error encountered while parsing references.

UnboundPortKind

The kind of unbound port.

Functions

is_valid_ident

Determines whether the provided token is a valid hyperactor identifier.

Type Aliases

Index

Index is a type alias representing a value that can be used as an index into a sequence.