hyperactor::reference

Struct ActorRef

pub struct ActorRef<A: Referable> { /* private fields */ }

ActorRefs are typed references to actors.

Implementations

impl<A: Referable> ActorRef<A>

pub fn port<M: RemoteMessage>(&self) -> PortRef<M>
where A: RemoteHandles<M>,

Get the remote port for message type [M] for the referenced actor.

pub fn send<M: RemoteMessage>( &self, cx: &impl Actor, message: M, ) -> Result<(), MailboxSenderError>
where A: RemoteHandles<M>,

Send an [M]-typed message to the referenced actor.

pub fn send_with_headers<M: RemoteMessage>( &self, cx: &impl Actor, headers: Flattrs, message: M, ) -> Result<(), MailboxSenderError>
where A: RemoteHandles<M>,

Send an [M]-typed message to the referenced actor, with additional context provided by headers.

pub fn attest(actor_id: ActorId) -> Self

The caller guarantees that the provided actor ID is also a valid, typed reference. This is usually invoked to provide a guarantee that an externally-provided actor ID (e.g., through a command line argument) is a valid reference.

pub fn actor_id(&self) -> &ActorId

The actor ID corresponding with this reference.

pub fn into_actor_id(self) -> ActorId

Convert this actor reference into its corresponding actor ID.

pub fn downcast_handle(&self, cx: &impl Actor) -> Option<ActorHandle<A>>
where A: Actor,

Attempt to downcast this reference into a (local) actor handle. This will only succeed when the referenced actor is in the same proc as the caller.