hyperactor

Module subject

Log subject for structured logging.

Each log message in hyperactor can have a subject: the entity the message pertains to (an actor, a proc, a channel, etc.). The subject is set as a tracing span field and rendered prominently in log output.

<actor metatls:host:1234,local,my_actor[0]> undeliverable message, ...

Structs

Subject

Identifies the entity a log message pertains to.

Traits

AsSubject

Extension trait for types that can be used as a log subject.