hyperactor::testing::proc_supervison

Struct ProcSupervisionCoordinator

pub struct ProcSupervisionCoordinator { /* private fields */ }

Used to create a proc supervison coordinator for testing purposes. Normally you should not use this struct. It is only required in the following cases:

The tests’ logic involves actor failures;

A supervison coordinator is not already set for the proc (e.g. the ProcActor scenario which will be explained later.)

This is because hyperactor’s supervision logic requires actor failures in a proc to be bubbled up to through the supervision chain:

grandchild actor -> child actor -> root actor -> proc supervison coordinator

If the the proc supervison coordinator is not set, supervision will crash the process because it cannot find the coordinator during the “bubbling up”.

Implementations

impl ProcSupervisionCoordinator

pub async fn set( proc: &Proc, ) -> Result<(ReportedEvent, ActorHandle<ProcSupervisionCoordinator>), Error>

Spawn a coordinator actor and set it as the coordinator for the given proc. Returns the reported event state and the coordinator actor handle. Callers should drain and stop the coordinator handle when the test is done to avoid ASAN thread-leak errors.