hyperactor::context

Trait Actor

pub trait Actor: Mailbox {
    type A: Actor;

    // Required method
    fn instance(&self) -> &Instance<Self::A>;
}

A typed actor context, providing both a Mailbox and an Instance.

Note: Send and Sync markers are here only temporarily in order to bridge the transition to the context types, away from the [crate::cap] module.

Required Associated Types

type A: Actor

The type of actor associated with this context.

Required Methods

fn instance(&self) -> &Instance<Self::A>

The instance associated with this context.