hyperactor

Module proc

This module provides Proc, which is the runtime used within a single proc.

Structs

ActorLedgerSnapshot

A snapshot view of the proc’s actor ledger.

ActorStats

Operational stats for an actor instance.

ActorTreeSnapshot

A snapshot of an actor tree (rooted at a pid=0 actor).

Context

Context for a message currently being handled by an Instance.

Event

A event for one row of log.

Instance

An actor instance. This is responsible for managing a running actor, including its full lifecycle, supervision, signal management, etc. Instances can represent a managed actor or a “client” actor that has joined the proc.

InstanceCell

InstanceCell contains all of the type-erased, shareable state of an instance. Specifically, InstanceCells form a supervision tree, and is used by ActorHandle to access the underlying instance.

Ports

A polymorphic dictionary that stores ports for an actor’s handlers. The interface memoizes the ports so that they are reused. We do not (yet) support stable identifiers across multiple instances of the same actor.

Proc

A proc instance is the runtime managing a single proc in Hyperactor. It is responsible for spawning actors in the proc, multiplexing messages to/within actors in the proc, and providing fallback routing to external procs.

SeqInfo

A message’s sequencer number infomation.

WeakInstanceCell

A weak version of the InstanceCell. This is used to provide cyclical linkage between actors without creating a strong reference cycle.

WorkCell

Represents a single work item used by the instance to dispatch to actor handles. Specifically, this enables handler polymorphism.

Statics

SEQ_INFO

The sender of this message, the session ID, and the message’s sequence number assigned by this session.