hyperactor::proc

Struct Proc

pub struct Proc { /* private fields */ }

A proc instance is the runtime managing a single proc in Hyperactor. It is responsible for spawning actors in the proc, multiplexing messages to/within actors in the proc, and providing fallback routing to external procs.

Procs are also responsible for maintaining the local supervision hierarchy.

Implementations

impl Proc

pub fn new(proc_id: ProcId, forwarder: BoxedMailboxSender) -> Self

Create a new proc with the given proc id and forwarder.

pub fn direct(addr: ChannelAddr, name: String) -> Result<Self, ChannelError>

Create a new direct-addressed proc.

pub fn direct_with_default( addr: ChannelAddr, name: String, default: BoxedMailboxSender, ) -> Result<Self, ChannelError>

Create a new direct-addressed proc with a default sender for the forwarder.

pub fn new_with_clock( proc_id: ProcId, forwarder: BoxedMailboxSender, clock: ClockKind, ) -> Self

Create a new proc with the given proc id, forwarder and clock kind.

pub fn set_supervision_coordinator( &self, port: PortHandle<ActorSupervisionEvent>, ) -> Result<(), Error>

Set the supervision coordinator’s port for this proc. Return Err if it is already set.

pub fn handle_unhandled_supervision_event( &self, cx: &impl Actor, event: ActorSupervisionEvent, )

Handle a supervision event received by the proc. Attempt to forward it to the supervision coordinator port if one is set, otherwise crash the process.

pub fn local() -> Self

Create a new local-only proc. This proc is not allowed to forward messages outside of the proc itself.

pub fn proc_id(&self) -> &ProcId

The proc’s ID.

pub fn forwarder(&self) -> &BoxedMailboxSender

Shared sender used by the proc to forward messages to remote destinations.

pub fn clock(&self) -> &ClockKind

The proc’s clock.

pub fn ledger_snapshot(&self) -> ActorLedgerSnapshot

Get the snapshot of the ledger.

pub fn attach(&self, name: &str) -> Result<Mailbox, Error>

Attach a mailbox to the proc with the provided root name.

pub fn attach_child(&self, parent_id: &ActorId) -> Result<Mailbox, Error>

Attach a mailbox to the proc as a child actor.

pub fn attach_actor<R, M>( &self, name: &str, ) -> Result<(Instance<()>, ActorRef<R>, PortReceiver<M>), Error>
where M: RemoteMessage, R: Referable + RemoteHandles<M>,

Attach a mailbox to the proc with the provided root name, and bind an ActorRef. This is intended only for testing, and will be replaced by simpled utilities.

pub fn spawn<A: Actor>( &self, name: &str, actor: A, ) -> Result<ActorHandle<A>, Error>

Spawn a named (root) actor on this proc. The name of the actor must be unique.

pub fn instance( &self, name: &str, ) -> Result<(Instance<()>, ActorHandle<()>), Error>

Wrapper for Proc::actor_instance::<()>.

pub fn actor_instance<A: Actor>( &self, name: &str, ) -> Result<(Instance<A>, ActorHandle<A>, PortReceiver<ActorSupervisionEvent>, PortReceiver<Signal>, UnboundedReceiver<WorkCell<A>>), Error>

Create and return an actor instance, its corresponding handle, its signal port receiver, its supervision port receiver, and its message receiver. This allows actors to be “inverted”: the caller can use the returned Instance to send and receive messages, launch child actors, etc. The actor itself does not handle any messages unless driven by the caller. Otherwise the instance acts as a normal actor, and can be referenced and stopped.

pub fn abort_root_actor( &self, root: &ActorId, this_handle: Option<&JoinHandle<()>>, ) -> Option<impl Future<Output = ActorId>>

Call abort on the JoinHandle associated with the given root actor. If successful return Some(root.clone()) else None.

pub fn stop_actor( &self, actor_id: &ActorId, reason: String, ) -> Option<Receiver<ActorStatus>>

Signals to a root actor to stop, returning a status observer if successful.

pub async fn destroy_and_wait<A: Actor>( &mut self, timeout: Duration, cx: Option<&Context<'_, A>>, reason: &str, ) -> Result<(Vec<ActorId>, Vec<ActorId>), Error>

Stop the proc. Returns a pair of:

the actors observed to stop;

the actors not observed to stop when timeout.

If cx is specified, it means this method was called from inside an actor in which case we shouldn’t wait for it to stop and need to delay aborting its task.

pub async fn destroy_and_wait_except_current<A: Actor>( &mut self, timeout: Duration, cx: Option<&Context<'_, A>>, except_current: bool, reason: &str, ) -> Result<(Vec<ActorId>, Vec<ActorId>), Error>

Stop the proc. Returns a pair of:

the actors observed to stop;

the actors not observed to stop when timeout.

If cx is specified, it means this method was called from inside an actor in which case we shouldn’t wait for it to stop and need to delay aborting its task. If except_current is true, don’t stop the actor represented by “cx” at all.

pub fn resolve_actor_ref<R: Actor + Referable>( &self, actor_ref: &ActorRef<R>, ) -> Option<ActorHandle<R>>

Resolve an actor reference to an actor residing on this proc. Returns None if the actor is not found on this proc.

Bounds:

R: Actor — must be a real actor that can live in this proc.

R: Referable — required because the input is an ActorRef<R>.